Move GetoptsState definition to yash-builtin and store it in Env::any
This commit moves the GetoptsState and Origin definitions from
yash-env to yash-builtin. The GetoptsState is stored in the Env::any
field, which is a DataSet. This change improves cohesion by keeping
the GetoptsState and Origin definitions inside the getopts module in the
yash-builtin crate.
