+++ yash-env/src/builtin.rs
@@ -31,8 +31,6 @@ use std::fmt::Debug;
 use std::future::Future;
 use std::pin::Pin;
 
-pub mod getopts;
-
 /// Types of built-in utilities
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum Type {
