+++ yash-env/src/lib.rs
@@ -32,7 +32,6 @@
 //! system's behavior without affecting the actual system.
 
 use self::any::DataSet;
-use self::builtin::getopts::GetoptsState;
 use self::builtin::Builtin;
 use self::function::FunctionSet;
 use self::io::Fd;
@@ -103,9 +102,6 @@ pub struct Env {
     /// Functions defined in the environment
     pub functions: FunctionSet,
 
-    /// State of the previous invocation of the `getopts` built-in
-    pub getopts_state: Option<GetoptsState>,
-
     /// Jobs managed in the environment
     pub jobs: JobList,
 
@@ -156,7 +152,6 @@ impl Env {
             builtins: Default::default(),
             exit_status: Default::default(),
             functions: Default::default(),
-            getopts_state: Default::default(),
             jobs: Default::default(),
             main_pgid: system.getpgrp(),
             main_pid: system.getpid(),
@@ -189,7 +184,6 @@ impl Env {
             builtins: self.builtins.clone(),
             exit_status: self.exit_status,
             functions: self.functions.clone(),
-            getopts_state: self.getopts_state.clone(),
             jobs: self.jobs.clone(),
             main_pgid: self.main_pgid,
             main_pid: self.main_pid,
